async_iot_gateway/requirements.txt
fastapi
uvicorn[standard]
redis
aiomysql
python-dotenv
